+++ src/config/mod.rs
@@ -1,0 +1,314 @@
+use de::read::BincodeRead;
+use error::Result;
+use serde;
+use std::io::{Read, Write};
+use std::marker::PhantomData;
+
+pub(crate) use self::endian::BincodeByteOrder;
+pub(crate) use self::int::IntEncoding;
+pub(crate) use self::internal::*;
+pub(crate) use self::limit::SizeLimit;
+
+pub use self::endian::{BigEndian, LittleEndian, NativeEndian};
+pub use self::int::{FixintEncoding, VarintEncoding};
+pub use self::legacy::*;
+pub use self::limit::{Bounded, Infinite};
+
+mod endian;
+mod int;
+mod legacy;
+mod limit;
+
+/// The default options for bincode serialization/deserialization.
+///
+/// ### Defaults
+/// By default bincode will use little-endian encoding for multi-byte integers, and will not
+/// limit the number of serialized/deserialized bytes.
+#[derive(Copy, Clone)]
+pub struct DefaultOptions(Infinite);
+
+impl DefaultOptions {
+    /// Get a default configuration object.
+    ///
+    /// ### Default Configuration:
+    ///
+    /// | Byte limit | Endianness | Int Encoding |
+    /// |------------|------------|--------------|
+    /// | Unlimited  | Little     | Varint       |
+    pub fn new() -> DefaultOptions {
+        DefaultOptions(Infinite)
+    }
+}
+
+impl Default for DefaultOptions {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl InternalOptions for DefaultOptions {
+    type Limit = Infinite;
+    type Endian = LittleEndian;
+    type IntEncoding = VarintEncoding;
+
+    #[inline(always)]
+    fn limit(&mut self) -> &mut Infinite {
+        &mut self.0
+    }
+}
+
+/// A configuration builder trait whose options Bincode will use
+/// while serializing and deserializing.
+///
+/// ### Options
+/// Endianness: The endianness with which multi-byte integers will be read/written.  *default: little endian*
+///
+/// Limit: The maximum number of bytes that will be read/written in a bincode serialize/deserialize. *default: unlimited*
+///
+/// ### Byte Limit Details
+/// The purpose of byte-limiting is to prevent Denial-Of-Service attacks whereby malicious attackers get bincode
+/// deserialization to crash your process by allocating too much memory or keeping a connection open for too long.
+///
+/// When a byte limit is set, bincode will return `Err` on any deserialization that goes over the limit, or any
+/// serialization that goes over the limit.
+/// Sets the byte limit to be unlimited.
+/// This is the default.
+pub trait Options: InternalOptions + Sized {
+    /// Sets the byte limit to be unlimited.
+    /// This is the default.
+    fn with_no_limit(self) -> WithOtherLimit<Self, Infinite> {
+        WithOtherLimit::new(self, Infinite)
+    }
+
+    /// Sets the byte limit to `limit`.
+    fn with_limit(self, limit: u64) -> WithOtherLimit<Self, Bounded> {
+        WithOtherLimit::new(self, Bounded(limit))
+    }
+
+    /// Sets the endianness to little-endian
+    /// This is the default.
+    fn with_little_endian(self) -> WithOtherEndian<Self, LittleEndian> {
+        WithOtherEndian::new(self)
+    }
+
+    /// Sets the endianness to big-endian
+    fn with_big_endian(self) -> WithOtherEndian<Self, BigEndian> {
+        WithOtherEndian::new(self)
+    }
+
+    /// Sets the endianness to the the machine-native endianness
+    fn with_native_endian(self) -> WithOtherEndian<Self, NativeEndian> {
+        WithOtherEndian::new(self)
+    }
+
+    /// Sets the length encoding to varint
+    fn with_varint_encoding(self) -> WithOtherIntEncoding<Self, VarintEncoding> {
+        WithOtherIntEncoding::new(self)
+    }
+
+    /// Sets the length encoding to be fixed
+    fn with_fixint_encoding(self) -> WithOtherIntEncoding<Self, FixintEncoding> {
+        WithOtherIntEncoding::new(self)
+    }
+
+    /// Serializes a serializable object into a `Vec` of bytes using this configuration
+    #[inline(always)]
+    fn serialize<S: ?Sized + serde::Serialize>(self, t: &S) -> Result<Vec<u8>> {
+        ::internal::serialize(t, self)
+    }
+
+    /// Returns the size that an object would be if serialized using Bincode with this configuration
+    #[inline(always)]
+    fn serialized_size<T: ?Sized + serde::Serialize>(self, t: &T) -> Result<u64> {
+        ::internal::serialized_size(t, self)
+    }
+
+    /// Serializes an object directly into a `Writer` using this configuration
+    ///
+    /// If the serialization would take more bytes than allowed by the size limit, an error
+    /// is returned and *no bytes* will be written into the `Writer`
+    #[inline(always)]
+    fn serialize_into<W: Write, T: ?Sized + serde::Serialize>(self, w: W, t: &T) -> Result<()> {
+        ::internal::serialize_into(w, t, self)
+    }
+
+    /// Deserializes a slice of bytes into an instance of `T` using this configuration
+    #[inline(always)]
+    fn deserialize<'a, T: serde::Deserialize<'a>>(self, bytes: &'a [u8]) -> Result<T> {
+        ::internal::deserialize(bytes, self)
+    }
+
+    /// TODO: document
+    #[doc(hidden)]
+    #[inline(always)]
+    fn deserialize_in_place<'a, R, T>(self, reader: R, place: &mut T) -> Result<()>
+    where
+        R: BincodeRead<'a>,
+        T: serde::de::Deserialize<'a>,
+    {
+        ::internal::deserialize_in_place(reader, self, place)
+    }
+
+    /// Deserializes a slice of bytes with state `seed` using this configuration.
+    #[inline(always)]
+    fn deserialize_seed<'a, T: serde::de::DeserializeSeed<'a>>(
+        self,
+        seed: T,
+        bytes: &'a [u8],
+    ) -> Result<T::Value> {
+        ::internal::deserialize_seed(seed, bytes, self)
+    }
+
+    /// Deserializes an object directly from a `Read`er using this configuration
+    ///
+    /// If this returns an `Error`, `reader` may be in an invalid state.
+    #[inline(always)]
+    fn deserialize_from<R: Read, T: serde::de::DeserializeOwned>(self, reader: R) -> Result<T> {
+        ::internal::deserialize_from(reader, self)
+    }
+
+    /// Deserializes an object directly from a `Read`er with state `seed` using this configuration
+    ///
+    /// If this returns an `Error`, `reader` may be in an invalid state.
+    #[inline(always)]
+    fn deserialize_from_seed<'a, R: Read, T: serde::de::DeserializeSeed<'a>>(
+        self,
+        seed: T,
+        reader: R,
+    ) -> Result<T::Value> {
+        ::internal::deserialize_from_seed(seed, reader, self)
+    }
+
+    /// Deserializes an object from a custom `BincodeRead`er using the default configuration.
+    /// It is highly recommended to use `deserialize_from` unless you need to implement
+    /// `BincodeRead` for performance reasons.
+    ///
+    /// If this returns an `Error`, `reader` may be in an invalid state.
+    #[inline(always)]
+    fn deserialize_from_custom<'a, R: BincodeRead<'a>, T: serde::de::DeserializeOwned>(
+        self,
+        reader: R,
+    ) -> Result<T> {
+        ::internal::deserialize_from_custom(reader, self)
+    }
+
+    /// Deserializes an object from a custom `BincodeRead`er with state `seed` using the default
+    /// configuration. It is highly recommended to use `deserialize_from` unless you need to
+    /// implement `BincodeRead` for performance reasons.
+    ///
+    /// If this returns an `Error`, `reader` may be in an invalid state.
+    #[inline(always)]
+    fn deserialize_from_custom_seed<'a, R: BincodeRead<'a>, T: serde::de::DeserializeSeed<'a>>(
+        self,
+        seed: T,
+        reader: R,
+    ) -> Result<T::Value> {
+        ::internal::deserialize_from_custom_seed(seed, reader, self)
+    }
+}
+
+impl<T: InternalOptions> Options for T {}
+
+/// A configuration struct with a user-specified byte limit
+#[derive(Clone, Copy)]
+pub struct WithOtherLimit<O: Options, L: SizeLimit> {
+    _options: O,
+    pub(crate) new_limit: L,
+}
+
+/// A configuration struct with a user-specified endian order
+#[derive(Clone, Copy)]
+pub struct WithOtherEndian<O: Options, E: BincodeByteOrder> {
+    options: O,
+    _endian: PhantomData<E>,
+}
+
+/// A configuration struct with a user-specified length encoding
+pub struct WithOtherIntEncoding<O: Options, I: IntEncoding> {
+    options: O,
+    _length: PhantomData<I>,
+}
+
+impl<O: Options, L: SizeLimit> WithOtherLimit<O, L> {
+    #[inline(always)]
+    pub(crate) fn new(options: O, limit: L) -> WithOtherLimit<O, L> {
+        WithOtherLimit {
+            _options: options,
+            new_limit: limit,
+        }
+    }
+}
+
+impl<O: Options, E: BincodeByteOrder> WithOtherEndian<O, E> {
+    #[inline(always)]
+    pub(crate) fn new(options: O) -> WithOtherEndian<O, E> {
+        WithOtherEndian {
+            options,
+            _endian: PhantomData,
+        }
+    }
+}
+
+impl<O: Options, I: IntEncoding> WithOtherIntEncoding<O, I> {
+    #[inline(always)]
+    pub(crate) fn new(options: O) -> WithOtherIntEncoding<O, I> {
+        WithOtherIntEncoding {
+            options,
+            _length: PhantomData,
+        }
+    }
+}
+
+impl<O: Options, E: BincodeByteOrder + 'static> InternalOptions for WithOtherEndian<O, E> {
+    type Limit = O::Limit;
+    type Endian = E;
+    type IntEncoding = O::IntEncoding;
+
+    #[inline(always)]
+    fn limit(&mut self) -> &mut O::Limit {
+        self.options.limit()
+    }
+}
+
+impl<O: Options, L: SizeLimit + 'static> InternalOptions for WithOtherLimit<O, L> {
+    type Limit = L;
+    type Endian = O::Endian;
+    type IntEncoding = O::IntEncoding;
+
+    fn limit(&mut self) -> &mut L {
+        &mut self.new_limit
+    }
+}
+
+impl<O: Options, I: IntEncoding + 'static> InternalOptions for WithOtherIntEncoding<O, I> {
+    type Limit = O::Limit;
+    type Endian = O::Endian;
+    type IntEncoding = I;
+
+    fn limit(&mut self) -> &mut O::Limit {
+        self.options.limit()
+    }
+}
+
+mod internal {
+    use super::*;
+
+    pub trait InternalOptions {
+        type Limit: SizeLimit + 'static;
+        type Endian: BincodeByteOrder + 'static;
+        type IntEncoding: IntEncoding + 'static;
+
+        fn limit(&mut self) -> &mut Self::Limit;
+    }
+
+    impl<'a, O: InternalOptions> InternalOptions for &'a mut O {
+        type Limit = O::Limit;
+        type Endian = O::Endian;
+        type IntEncoding = O::IntEncoding;
+
+        #[inline(always)]
+        fn limit(&mut self) -> &mut Self::Limit {
+            (*self).limit()
+        }
+    }
+}
