+++ tests/test.rs
@@ -281,12 +281,14 @@ fn deserializing_errors() {
 fn too_big_deserialize() {
     let serialized = vec![0, 0, 0, 3];
     let deserialized: Result<u32> = DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(3)
         .deserialize_from(&mut &serialized[..]);
     assert!(deserialized.is_err());
 
     let serialized = vec![0, 0, 0, 3];
     let deserialized: Result<u32> = DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(4)
         .deserialize_from(&mut &serialized[..]);
     assert!(deserialized.is_ok());
@@ -318,16 +320,23 @@ fn too_big_char_deserialize() {
 #[test]
 fn too_big_serialize() {
     assert!(DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(3)
         .serialize(&0u32)
         .is_err());
-    assert!(DefaultOptions::new().with_limit(4).serialize(&0u32).is_ok());
+    assert!(DefaultOptions::new()
+        .with_fixint_encoding()
+        .with_limit(4)
+        .serialize(&0u32)
+        .is_ok());
 
     assert!(DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(LEN_SIZE + 4)
         .serialize(&"abcde")
         .is_err());
     assert!(DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(LEN_SIZE + 5)
         .serialize(&"abcde")
         .is_ok());
@@ -352,6 +361,7 @@ fn test_serialized_size_bounded() {
     // JUST RIGHT
     assert!(
         DefaultOptions::new()
+            .with_fixint_encoding()
             .with_limit(1)
             .serialized_size(&0u8)
             .unwrap()
@@ -359,6 +369,7 @@ fn test_serialized_size_bounded() {
     );
     assert!(
         DefaultOptions::new()
+            .with_fixint_encoding()
             .with_limit(2)
             .serialized_size(&0u16)
             .unwrap()
@@ -366,6 +377,7 @@ fn test_serialized_size_bounded() {
     );
     assert!(
         DefaultOptions::new()
+            .with_fixint_encoding()
             .with_limit(4)
             .serialized_size(&0u32)
             .unwrap()
@@ -373,6 +385,7 @@ fn test_serialized_size_bounded() {
     );
     assert!(
         DefaultOptions::new()
+            .with_fixint_encoding()
             .with_limit(8)
             .serialized_size(&0u64)
             .unwrap()
@@ -380,6 +393,7 @@ fn test_serialized_size_bounded() {
     );
     assert!(
         DefaultOptions::new()
+            .with_fixint_encoding()
             .with_limit(8)
             .serialized_size(&"")
             .unwrap()
@@ -387,6 +401,7 @@ fn test_serialized_size_bounded() {
     );
     assert!(
         DefaultOptions::new()
+            .with_fixint_encoding()
             .with_limit(8 + 1)
             .serialized_size(&"a")
             .unwrap()
@@ -394,6 +409,7 @@ fn test_serialized_size_bounded() {
     );
     assert!(
         DefaultOptions::new()
+            .with_fixint_encoding()
             .with_limit(LEN_SIZE + 3 * 4)
             .serialized_size(&vec![0u32, 1u32, 2u32])
             .unwrap()
@@ -401,30 +417,37 @@ fn test_serialized_size_bounded() {
     );
     // Below
     assert!(DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(0)
         .serialized_size(&0u8)
         .is_err());
     assert!(DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(1)
         .serialized_size(&0u16)
         .is_err());
     assert!(DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(3)
         .serialized_size(&0u32)
         .is_err());
     assert!(DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(7)
         .serialized_size(&0u64)
         .is_err());
     assert!(DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(7)
         .serialized_size(&"")
         .is_err());
     assert!(DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(8 + 0)
         .serialized_size(&"a")
         .is_err());
     assert!(DefaultOptions::new()
+        .with_fixint_encoding()
         .with_limit(8 + 3 * 4 - 1)
         .serialized_size(&vec![0u32, 1u32, 2u32])
         .is_err());
